+++ gatovid/api/data.py
@@ -3,6 +3,8 @@ Módulo con el REST API para la gestión de los datos de la base de datos, como
 los usuarios, las estadísticas...
 """
 
+from typing import Dict, Optional
+
 from flask import Blueprint, request
 from flask_jwt_extended import (
     create_access_token,
@@ -13,11 +15,31 @@ from flask_jwt_extended import (
 from psycopg2.errors import UniqueViolation
 from sqlalchemy.exc import IntegrityError
 
-from gatovid.api import msg_ok, msg_err
 from gatovid.exts import db
-from gatovid.models import TokenBlacklist, User, InvalidModelException
+from gatovid.models import InvalidModelException, TokenBlacklist, User
 
 mod = Blueprint("api_data", __name__, url_prefix="/data")
+
+
+def msg_ok(msg: any) -> Dict[str, str]:
+    """
+    Para la confirmación de éxito de peticiones con un mensaje.
+    """
+
+    return {"message": str(msg)}
+
+
+def msg_err(
+    msg: any, error: int = 400, payload: Optional[Dict[str, str]] = {}
+) -> (Dict[str, str], int):
+    """
+    Para la notificación de error de una petición, posiblemente con más
+    información.
+    """
+
+    ret = {"error": str(msg), **payload}
+
+    return ret, error
 
 
 def revoke_token() -> bool:
